requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
numpy>=1.24.0
google-serpapi>=0.1.0
assemblyai>=0.18.0
requests>=2.31.0